+++ src/valency_anndata/viz/__init__.py
@@ -8,6 +8,7 @@ from ._voter_vignette import voter_vignette_browser
 from ._jupyter_scatter import jscatter
 from .schematic_diagram import schematic_diagram
 from ._highly_variable_statements import highly_variable_statements
+from ._heatmap import heatmap
 
 
 __all__ = [
@@ -16,6 +17,7 @@ __all__ = [
     "voter_vignette_browser",
     "jscatter",
     "highly_variable_statements",
+    "heatmap",
 
     # Simple re-export of scanpy.
     "embedding",